requirements.txt
streamlit
google-genai
google-cloud-logging
google-cloud-aiplatform
pandas
